+++ CHANGELOG.md
@@ -15,7 +15,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - 決済サービスプロバイダーのIPaymentUrlSettingsにuse3DSを追加
 - 注文オファーにserialNumberを追加
 - オファー承認アクションにinstrument.transactionNumberを保証
-- IAMメンバーにsettingsを追加
+- 販売者のmakesOfferにeligibleMembershipを追加
 
 ### Changed